add: repository 계층 구현

+++ back/src/main/java/com/oneorthree/phone/repository/CharacterEquipmentRepository.java
@@ -2,13 +2,13 @@ package com.oneorthree.phone.repository;
 
 import com.oneorthree.phone.domain.CharacterEquipment;
 import com.oneorthree.phone.domain.SlotType;
+import com.oneorthree.phone.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.security.core.userdetails.User;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface CharacterEquipmentRepository extends JpaRepository<CharacterEquipment, Integer> {
+public interface CharacterEquipmentRepository extends JpaRepository<CharacterEquipment, Long> {
     // 유저의 전체 장착 상태 조회 (캐릭터 렌더링)
     List<CharacterEquipment> findByUser(User user);
